Move Controller to namespace `Vaccine`
